fix embarrassing bug of passing the wrong DF!
I accidentally passed in the wrong DF into the functions that perform
the fits. This resulted in too much data in each fit and thus a very
bad fit in pretty much all cases.

Sigh.

I.e. the data contained not just the Fibonacci benchmarks, but also
Keccak etc, which don't belong in there. In addition in every case it
included not just the serial (or parallel) benchmark, but also the
parallel (or serial).