restructure fitting & facet plot logic
Instead of mixing the two, we now first perform all fits and store
everything in a `Table[Class, FitResult]`. From this we can
reconstruct a DF that contains everything we need (including the
yMin,yMax values for error bars).

This allows us to easily produce facet plots for either all metrics
_or_ all traces.